+++ tests/testthat/test-helpers.R
@@ -43,23 +43,23 @@ test_that("edc_warn_patient_diffs works", {
 
 
 
-test_that("harmonize_subjid works", {
+test_that("edc_unify_subjid works", {
   #numeric subjid
   db = list(x=data.frame(subjid=c("1", "2", "3", "005"), a=1))
-  h = harmonize_subjid(db, col_subjid="subjid")
+  h = edc_unify_subjid(db, col_subjid="subjid")
   expect_identical(h$x$subjid, factor(c("1", "2", "3", "5")))
-  h = harmonize_subjid(db, col_subjid="subjid", preprocess=identity)
+  h = edc_unify_subjid(db, col_subjid="subjid", preprocess=identity)
   expect_identical(h$x$subjid, factor(c("1", "2", "3", "005"), 
                                       levels = c("1", "2", "3", "005")))
   
   #factor subjid
   db = list(x=data.frame(subjid=as_factor(c("1", "2", "7", "5")), a=1))
-  h = harmonize_subjid(db, col_subjid="subjid")
+  h = edc_unify_subjid(db, col_subjid="subjid")
   expect_identical(h$x$subjid, factor(c("1", "2", "7", "5")))
   
   #character subjid
   db = list(x=data.frame(subjid=c("pat-1", "pat-3", "pat-2"), a=1))
-  h = harmonize_subjid(db, col_subjid="subjid")
+  h = edc_unify_subjid(db, col_subjid="subjid")
   expect_identical(h$x$subjid, factor(c("pat-1", "pat-3", "pat-2")))
 })